New function graphql-simplify-response-edges
GraphQL APIs will return networked objects in terms of 'edges' and
'nodes'.  This function detects those relationships in a response data
structure and collapses them into just the value of each 'node'.

For instance, if I had a data structure like this:

    (some_object (edges ((node (prop . "value")))
                        ((node (prop . "other")))))

the simplification would look like this:

    (some_object ((prop . "value"))
                 ((prop . "other"))

Of course, this will work recursively and descend into fields and
subfields to simplify all edges it finds.

This does have the side-effect of removing pagination utilities like
the 'cursor' field of an edge, but that is a complex problem all its
own and it's probably best if this information is retrieved by other
means.